# contracts/openapi/wiki-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nova Haven Content API",
    "version": "1.0.0",
    "description": "Draft/public isolation for Wiki, News and Catalog: anonymous clients only see published revisions. Editorial operations require Admin cookie + CSRF; mutation preconditions use If-Match."
  },
  "paths": {
...
    "/api/v1/catalog/items": {
      "get": {
        "tags": ["Catalog"],
        "parameters": [
          {"name": "q", "in": "query", "schema": {"type": "string", "maxLength": 100}},
          {"name": "kind", "in": "query", "schema": {"$ref": "#/components/schemas/CatalogItemKind"}},
          {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50}}
        ],
        "responses": {"200": {"description": "Published catalog item snapshots only"}, "400": {"description": "Invalid filter"}}
      }
    },
    "/api/v1/catalog/items/{slug}": {
      "get": {
        "tags": ["Catalog"],
        "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Published catalog item snapshot"}, "404": {"description": "Unknown or unpublished item"}}
      }
    },
    "/api/v1/catalog/recipes": {
      "get": {
        "tags": ["Catalog"],
        "parameters": [
          {"name": "q", "in": "query", "schema": {"type": "string", "maxLength": 100}},
          {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50}}
        ],
        "responses": {"200": {"description": "Published recipe snapshots only"}, "400": {"description": "Invalid filter"}}
      }
    },
    "/api/v1/catalog/recipes/{slug}": {
      "get": {
        "tags": ["Catalog"],
        "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Published recipe with immutable item revision components"}, "404": {"description": "Unknown or unpublished recipe"}}
      }
    },
...
    "/api/v1/admin/catalog/items/{id}": {
      "get": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "Catalog draft with ETag"}, "404": {"description": "Unknown item"}}},
      "patch": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"name": "If-Match", "in": "header", "required": true, "schema": {"type": "string"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CatalogItemInput"}}}}, "responses": {"200": {"description": "Updated with ETag"}, "412": {"description": "Stale ETag"}, "428": {"description": "Missing ETag"}}}
    },
...
    "/api/v1/admin/catalog/recipes/{id}": {
      "get": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "Recipe draft with ETag"}, "404": {"description": "Unknown recipe"}}},
      "patch": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"name": "If-Match", "in": "header", "required": true, "schema": {"type": "string"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CatalogRecipeInput"}}}}, "responses": {"200": {"description": "Updated with ETag"}, "412": {"description": "Stale ETag"}, "428": {"description": "Missing ETag"}}}
    },
...
    "/api/v1/community": {
      "get": {"tags": ["Community"], "parameters": [{"name": "kind", "in": "query", "schema": {"$ref": "#/components/schemas/CommunityKind"}}, {"name": "q", "in": "query", "schema": {"type": "string", "maxLength": 100}}, {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1}}, {"name": "pageSize", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50}}], "responses": {"200": {"description": "Published community records only"}}}
    },
    "/api/v1/community/{kind}": {
      "get": {"tags": ["Community"], "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/CommunityKind"}}], "responses": {"200": {"description": "Published records of the selected kind"}, "400": {"description": "Invalid kind"}}}
    },
    "/api/v1/community/{kind}/{slug}": {
      "get": {"tags": ["Community"], "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/CommunityKind"}}, {"name": "slug", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Published community record"}, "404": {"description": "Unknown or unpublished record"}}}
    },
    "/api/v1/admin/community/{kind}": {
      "get": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/CommunityKind"}}], "responses": {"200": {"description": "Community drafts"}}},
      "post": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/CommunityKind"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommunityInput"}}}}, "responses": {"201": {"description": "Draft created"}, "400": {"description": "Validation"}, "409": {"description": "Duplicate slug"}}}
    },
    "/api/v1/admin/community/{kind}/{id}": {
      "get": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/CommunityKind"}}, {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "Community draft with ETag"}}},
      "patch": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/CommunityKind"}}, {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"name": "If-Match", "in": "header", "required": true, "schema": {"type": "string"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommunityInput"}}}}, "responses": {"200": {"description": "Updated with ETag"}, "412": {"description": "Stale ETag"}, "428": {"description": "Missing ETag"}}}
    },
...
    "/api/v1/commerce/offers": {
      "get": {"tags": ["Commerce"], "parameters": [{"name": "q", "in": "query", "schema": {"type": "string", "maxLength": 100}}, {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1}}, {"name": "pageSize", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50}}], "responses": {"200": {"description": "Published offers and their published local-demo purchase flags/prices", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommerceOfferPage"}}}}}}
    },
    "/api/v1/commerce/offers/{slug}": {
      "get": {"tags": ["Commerce"], "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Published commerce offer", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommerceOfferDetail"}}}}, "404": {"description": "Unknown or unpublished offer"}}}
    },
    "/api/v1/commerce/orders": {
      "post": {"tags": ["Commerce"], "parameters": [{"name": "Idempotency-Key", "in": "header", "required": true, "description": "UUID reused for safe retry of the same cart.", "schema": {"type": "string", "format": "uuid"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommerceCheckoutRequest"}}}}, "responses": {"201": {"description": "Local demo order stored; no real charge or game fulfillment", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LocalDemoCheckoutReceipt"}}}}, "200": {"description": "Idempotent replay of the original local demo order", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LocalDemoCheckoutReceipt"}}}}, "400": {"description": "Invalid cart or missing idempotency key", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}}, "409": {"description": "Offer unavailable or idempotency key reused for a different cart", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}}}}
    },
...
    "/api/v1/admin/commerce/offers/{id}": {
      "get": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "Commerce draft with ETag"}}},
      "patch": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"name": "If-Match", "in": "header", "required": true, "schema": {"type": "string"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommerceInput"}}}}, "responses": {"200": {"description": "Updated with ETag"}, "412": {"description": "Stale definition"}, "428": {"description": "Missing ETag"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": ".AspNetCore.Identity.Application"
      },
      "userCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": ".AspNetCore.Identity.Application"
      }
    },
    "schemas": {
...
      "CommunityKind": {"type": "string", "enum": ["event", "guild", "player", "housing", "leaderboard"]},
      "CommunityLeaderboardRow": {
        "type": "object", "additionalProperties": false, "required": ["rank", "participantName", "score"],
        "properties": {"rank": {"type": "integer", "minimum": 1}, "participantName": {"type": "string", "minLength": 1, "maxLength": 120}, "score": {"type": "number"}, "note": {"type": "string", "maxLength": 300}}
      },
      "CommunityInput": {
        "type": "object", "additionalProperties": false, "required": ["name", "slug", "summary", "markdown"],
        "properties": {
          "name": {"type": "string", "minLength": 2, "maxLength": 160}, "slug": {"type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "maxLength": 160}, "summary": {"type": "string", "maxLength": 500}, "markdown": {"type": "string", "minLength": 1, "maxLength": 100000}, "kind": {"$ref": "#/components/schemas/CommunityKind"},
          "startsAt": {"type": ["string", "null"], "format": "date-time"}, "endsAt": {"type": ["string", "null"], "format": "date-time"}, "locationEntryId": {"type": ["string", "null"], "format": "uuid"}, "capacity": {"type": ["integer", "null"], "minimum": 1, "maximum": 100000}, "registrationOpen": {"type": "boolean"}, "motto": {"type": "string", "maxLength": 300}, "discordUrl": {"type": "string", "maxLength": 500}, "handle": {"type": "string", "maxLength": 80}, "bio": {"type": "string", "maxLength": 2000}, "avatarUrl": {"type": "string", "maxLength": 500}, "ownerDisplayName": {"type": "string", "maxLength": 120}, "galleryMarkdown": {"type": "string", "maxLength": 10000}, "leaderboardCategory": {"type": "string", "maxLength": 120}, "seasonEntryId": {"type": ["string", "null"], "format": "uuid"}, "rows": {"type": "array", "maxItems": 100, "items": {"$ref": "#/components/schemas/CommunityLeaderboardRow"}}
        }
      },
...
      "CommerceKind": {"type": "string", "enum": ["cosmetic", "membership", "bundle", "donation", "other"]},
      "CommerceInput": {"type": "object", "additionalProperties": false, "required": ["name", "slug", "summary", "markdown", "kind", "displayPrice"], "properties": {"name": {"type": "string", "maxLength": 160}, "slug": {"type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "maxLength": 120}, "summary": {"type": "string", "maxLength": 500}, "markdown": {"type": "string", "maxLength": 50000}, "kind": {"$ref": "#/components/schemas/CommerceKind"}, "displayPrice": {"type": "string", "maxLength": 80}, "providerProductCode": {"type": ["string", "null"], "maxLength": 120}, "isPurchasable": {"type": "boolean", "default": false}, "priceMinorUnits": {"type": ["integer", "null"], "minimum": 0, "maximum": 1000000000000, "description": "Authoritative numeric VND price; required and positive when isPurchasable is true."}}},
      "CommerceOfferSummary": {"type": "object", "required": ["id", "slug", "name", "summary", "kind", "revision", "definitionOnly", "isPurchasable", "priceMinorUnits", "currencyCode", "checkoutMode", "updatedAt"], "properties": {"id": {"type": "string", "format": "uuid"}, "slug": {"type": "string"}, "name": {"type": "string"}, "summary": {"type": "string"}, "kind": {"$ref": "#/components/schemas/CommerceKind"}, "revision": {"type": "integer"}, "definitionOnly": {"type": "boolean", "const": true}, "isPurchasable": {"type": "boolean"}, "priceMinorUnits": {"type": ["integer", "null"]}, "currencyCode": {"type": "string", "const": "VND"}, "checkoutMode": {"type": ["string", "null"], "enum": ["localDemo", null]}, "updatedAt": {"type": "string", "format": "date-time"}}},
      "CommerceOfferPage": {"type": "object", "required": ["items", "page", "pageSize", "total"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/CommerceOfferSummary"}}, "page": {"type": "integer"}, "pageSize": {"type": "integer"}, "total": {"type": "integer"}}},
      "CommerceOfferDetail": {"allOf": [{"$ref": "#/components/schemas/CommerceOfferSummary"}, {"type": "object", "required": ["markdown", "displayPrice", "publishedAt"], "properties": {"markdown": {"type": "string"}, "displayPrice": {"type": "string"}, "providerProductCode": {"type": ["string", "null"]}, "publishedAt": {"type": "string", "format": "date-time"}}}]},
      "CommerceCheckoutLineInput": {"type": "object", "required": ["slug", "quantity"], "properties": {"slug": {"type": "string", "minLength": 3, "maxLength": 120, "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"}, "quantity": {"type": "integer", "minimum": 1, "maximum": 99}}},
      "CommerceCheckoutRequest": {"type": "object", "required": ["items"], "properties": {"items": {"type": "array", "description": "Contains no duplicate slug values.", "minItems": 1, "maxItems": 20, "items": {"$ref": "#/components/schemas/CommerceCheckoutLineInput"}}}},
      "CommerceOrderLine": {"type": "object", "required": ["slug", "name", "revision", "quantity", "unitPriceMinorUnits", "lineTotalMinorUnits"], "properties": {"slug": {"type": "string"}, "name": {"type": "string"}, "revision": {"type": "integer"}, "quantity": {"type": "integer", "minimum": 1, "maximum": 99}, "unitPriceMinorUnits": {"type": "integer", "minimum": 1}, "lineTotalMinorUnits": {"type": "integer", "minimum": 1}}},
      "LocalDemoCheckoutReceipt": {"type": "object", "required": ["orderNumber", "createdAt", "status", "totalMinorUnits", "currencyCode", "items", "paymentStatus", "paymentMethod", "realCharge", "fulfilment", "idempotentReplay", "message"], "properties": {"orderNumber": {"type": "string"}, "createdAt": {"type": "string", "format": "date-time"}, "status": {"type": "string", "const": "demoCompleted"}, "totalMinorUnits": {"type": "integer", "minimum": 1}, "currencyCode": {"type": "string", "const": "VND"}, "items": {"type": "array", "items": {"$ref": "#/components/schemas/CommerceOrderLine"}}, "paymentStatus": {"type": "string", "const": "simulated"}, "paymentMethod": {"type": "string", "const": "localDemo"}, "realCharge": {"type": "boolean", "const": false}, "fulfilment": {"type": "string", "const": "none"}, "idempotentReplay": {"type": "boolean"}, "message": {"type": "string"}}},
...
      "ProblemDetails": {"type": "object", "properties": {"type": {"type": "string"}, "title": {"type": "string"}, "status": {"type": "integer"}, "detail": {"type": "string"}, "instance": {"type": "string"}}},
      "CatalogItemKind": {
        "type": "string",
        "enum": ["item", "weapon", "armor", "material", "fish", "crop"]
      },
      "CatalogItemInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "slug", "summary", "markdown", "kind"],
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 160},
          "slug": {"type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "minLength": 3, "maxLength": 120},
          "summary": {"type": "string", "maxLength": 500},
          "markdown": {"type": "string", "minLength": 1, "maxLength": 50000},
          "kind": {"$ref": "#/components/schemas/CatalogItemKind"}
        }
      },
      "CatalogRecipeComponent": {
        "type": "object",
        "additionalProperties": false,
        "required": ["itemId", "quantity"],
        "properties": {
          "itemId": {"type": "string", "format": "uuid"},
          "quantity": {"type": "integer", "minimum": 1, "maximum": 999}
        }
      },
      "CatalogRecipeInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "slug", "summary", "markdown", "ingredients", "outputs"],
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 160},
          "slug": {"type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "minLength": 3, "maxLength": 120},
          "summary": {"type": "string", "maxLength": 500},
          "markdown": {"type": "string", "minLength": 1, "maxLength": 50000},
          "ingredients": {"type": "array", "minItems": 1, "maxItems": 16, "items": {"$ref": "#/components/schemas/CatalogRecipeComponent"}},
          "outputs": {"type": "array", "minItems": 1, "maxItems": 8, "items": {"$ref": "#/components/schemas/CatalogRecipeComponent"}}
        }
      },
...
    }
  }
}